remove entry-field, modify search-input
