--- packages/vue-component-library/src/stories/EntryField.spec.js
@@ -1,7 +1,0 @@
-describe('Funkhaus / Entry Field', () => {
-  it('Default', () => {
-    cy.visit('/iframe.html?id=funkhaus-entry-field--default&args=&viewMode=story')
-    cy.get('.entry-field').should('exist')
-    cy.percySnapshot('Funkhaus / Entry Field')
-  })
-})
