+++ packages/vue-component-library/src/stories/SearchInput.spec.js
@@ -1,0 +1,9 @@
+describe('Funkhaus / Search Input', () => {
+  it('Default', () => {
+    cy.visit(
+      '/iframe.html?id=funkhaus-search-input--default&args=&viewMode=story'
+    )
+    cy.get('.search-input').should('exist')
+    cy.percySnapshot('Funkhaus / Search Input')
+  })
+})
