+++ packages/vue-component-library/src/styles/default/_search-input.scss
@@ -1,0 +1,98 @@
+$input-color: #333;
+$input-background: #f6f9fc;
+$icon-color: darken($input-background, 30%);
+$active-color: #1ea7fd;
+
+.search-input-wrapper {
+    position: relative;
+
+    input[data-search-input="true"] {
+        display: block;
+        font-family: var(--font-primary);
+        font-style: normal;
+        font-weight: normal;
+        font-size: 20px;
+        line-height: 100%;
+        letter-spacing: 0.01em;
+        background-color: var(--color-primary-blue-01);
+        border-color: transparent;
+        padding: 24px 24px 24px 16px;
+        width: 100%;
+
+        &::placeholder {
+            text-transform: uppercase;
+            font-family: var(--font-primary);
+            text-overflow: ellipsis;
+        }
+
+        /*
+      &:focus {
+          background-color: var(--color-primary-blue-01);
+          border-color: $active-color;
+          outline: 0;
+          box-shadow: none;
+      }*/
+    }
+
+    .clear-icon {
+        color: $icon-color;
+        position: absolute;
+
+        right: 15px;
+        bottom: 22px;
+        cursor: pointer;
+        z-index: 10;
+        box-sizing: border-box;
+        display: block;
+        width: 24px;
+        height: 24px;
+        border: 2px solid transparent;
+        border-radius: 40px;
+        background: none;
+        padding: 0px;
+        outline: none;
+
+        &:focus {
+            background: darken($input-background, 4%);
+        }
+
+        &::after,
+        &::before {
+            content: "";
+            display: block;
+            box-sizing: border-box;
+            position: absolute;
+            width: 16px;
+            height: 2px;
+            background: $icon-color;
+            transform: rotate(45deg);
+            border-radius: 5px;
+            top: 9px;
+            left: 2px;
+        }
+
+        &::after {
+            transform: rotate(-45deg);
+        }
+    }
+}
+
+/* Fix the X appearing in search field on Chrome and IE */
+input[type="search"]::-ms-clear {
+    display: none;
+    width: 0;
+    height: 0;
+}
+
+input[type="search"]::-ms-reveal {
+    display: none;
+    width: 0;
+    height: 0;
+}
+
+input[type="search"]::-webkit-search-decoration,
+input[type="search"]::-webkit-search-cancel-button,
+input[type="search"]::-webkit-search-results-button,
+input[type="search"]::-webkit-search-results-decoration {
+    display: none;
+}
